requirements/base.txt
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

joblib>=1.0.0
tqdm>=4.62.0
pyyaml>=6.0
